refactor: remove unused UpdateRegulationDto and simplify validation error handling

+++ src/modules/regulation/regulation.service.ts
@@ -1,8 +1,4 @@
-import {
-  BadRequestException,
-  Injectable,
-  NotFoundException,
-} from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository, DataSource } from 'typeorm';
 import { CreateRegulationDto } from './dto/create-regulation.dto';
@@ -60,90 +56,74 @@ export class RegulationService {
     await queryRunner.startTransaction();
 
     try {
-      // Create the main regulation first
       const regulation = await queryRunner.manager.save(Regulation, data);
 
-      // Create levels
       await queryRunner.manager.save(
         Level,
         levels.map((level) => ({ ...level, regulation })),
       );
 
-      // Create cumGpaRanges
       await queryRunner.manager.save(
         CumGpaRange,
         cumGpaRanges.map((range) => ({ ...range, regulation })),
       );
 
-      // Create courseGpaRanges
       await queryRunner.manager.save(
         CourseGpaRange,
         courseGpaRanges.map((range) => ({ ...range, regulation })),
       );
 
-      // Create registrationRules
       await queryRunner.manager.save(RegistrationRules, {
         ...registrationRules,
         regulation,
       });
 
-      // Create academicRequirements
       await queryRunner.manager.save(AcademicRequirements, {
         ...academicRequirements,
         regulation,
       });
 
-      // Create universityRequirements
       await queryRunner.manager.save(UniversityRequirements, {
         ...universityRequirements,
         regulation,
       });
 
-      // Create specializationRequirements
       const specialization = await queryRunner.manager.save(
         SpecializationRequirements,
         { ...specData, regulation },
       );
 
-      // Create trainingRequirements
       await queryRunner.manager.save(TrainingRequirements, {
         ...trainingRequirements,
         specializationRequirements: specialization,
       });
 
-      // Create gradProjectRequirements
       await queryRunner.manager.save(GradProjectRequirements, {
         ...gradProjectRequirements,
         specializationRequirements: specialization,
       });
 
-      // Create facultyRequirements
       await queryRunner.manager.save(FacultyRequirements, {
         ...facultyRequirements,
         regulation,
       });
 
-      // Create basicScienceRequirements
       await queryRunner.manager.save(BasicScienceRequirements, {
         ...basicScienceRequirements,
         regulation,
       });
 
-      // Create retakeRules
       await queryRunner.manager.save(RetakeRules, {
         ...retakeRules,
         regulation,
       });
 
-      // Create dismissalRules
       await queryRunner.manager.save(DismissalRules, {
         ...dismissalRules,
         regulation,
       });
 
       await queryRunner.commitTransaction();
-
-      return this.findOne(regulation.id);
     } catch (err) {
       await queryRunner.rollbackTransaction();
       throw err;
@@ -314,7 +294,6 @@ export class RegulationService {
         universityRequirements,
       );
 
-      // Update specializationRequirements
       await queryRunner.manager.update(
         SpecializationRequirements,
         { regulation: { id } },
@@ -324,24 +303,16 @@ export class RegulationService {
       const spec = await queryRunner.manager.findOne(
         SpecializationRequirements,
         {
-          where: { regulation: { id: id } },
+          where: { regulation: { id } },
         },
       );
 
-      if (!spec) {
-        throw new BadRequestException(
-          ErrorEnum.SPECIALIZATION_REQUIREMENTS_NOT_FOUND,
-        );
-      }
-
-      // Update trainingRequirements
       await queryRunner.manager.update(
         TrainingRequirements,
         { specializationRequirements: { id: spec.id } },
         trainingRequirements,
       );
 
-      // Update gradProjectRequirements
       await queryRunner.manager.update(
         GradProjectRequirements,
         { specializationRequirements: { id: spec.id } },
@@ -374,8 +345,6 @@ export class RegulationService {
 
       await queryRunner.manager.update(Regulation, { id }, data);
       await queryRunner.commitTransaction();
-
-      return this.findOne(id);
     } catch (err) {
       await queryRunner.rollbackTransaction();
       throw err;
@@ -385,7 +354,10 @@ export class RegulationService {
   }
 
   async remove(id: Regulation['id']) {
-    const regulation = await this.findOne(id);
+    const regulation = await this.regulationRepo.findOne({ where: { id } });
+    if (!regulation) {
+      throw new NotFoundException(ErrorEnum.REGULATION_NOT_FOUND);
+    }
     return this.regulationRepo.remove(regulation);
   }
 }
